refactor tests files
